Arkanoid Brick
Well-known game app from the 1980s: Arkanoid Brick Break \
Game Bricks Bricks (Static) Moving Bricks (located at the top of the screen, a racket)
Paddle (located at the bottom of the screen and can move horizontally) left and right (at the player's decision,
And a ball (circle) that moves and bounces on the screen autonomously.
Purpose of the game - Break all the bricks using the on-screen moving ball. White breaks and disappears when the ball hits
in her. The ball bounces every time it hits the screen / brick / bat boundary. The player must avoid the ball
Comes to the bottom of the screen (under the bat), which causes loss of life. To avoid loss of life, the player must move
The bat to the right and left so that the ball bounces on it and continues to move upward, to allow the continuation
Game and goal achievement. The player gets a score every time he breaks a brick, but loses life every time the ball
Falling under the bat.
The game ends in one of two situations: in the win if the player hit all the bricks, in the loss if the player ended up alive.